Introduce a short-term cache of reused values during epoch processing (matching attestations, base rewards etc).
