+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/teku/reference/phase0/epoch_processing/EpochProcessingTestExecutor.java
@@ -17,7 +17,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static tech.pegasys.teku.reference.phase0.TestDataUtils.loadStateFromSsz;
 
 import com.google.common.collect.ImmutableMap;
-import tech.pegasys.teku.core.EpochProcessorUtil;
+import tech.pegasys.teku.core.epoch.EpochProcessorUtil;
 import tech.pegasys.teku.datastructures.state.BeaconState;
 import tech.pegasys.teku.datastructures.state.BeaconState.Mutator;
 import tech.pegasys.teku.ethtests.finder.TestDefinition;
